Extend NodeDB to allow use of an external store for leaf node values
This makes it possible to persist leaf node values separately from the
tree nodes themselves. As leaf nodes are loaded their values are read
from the external store, when they're persisted to disk the values are
omitted. The external value store needs to be updated separately when
a key is set or deleted in the MutableTree.