proof: revert to using Tendermint merkle package (#291)
In #270 we copied some types from Tendermint into IAVL. However, this won't work for the proofs. The Tendermint `ProofRuntime` requires registered proof decoders that return `merkle.ProofOperator` but the IAVL proof decoders return `iavl.ProofOperator` and thus cannot be used with the proof runtime.

This reverts to using the Tendermint `merkle` package.