grpc/proto changes for GetByIndex (#355)
#354 

Is there any documentation on how to do the setup for the gogoproto generation stuff? I see the `proto-gen` make action, but when I tried to set things up manually to do that I ended up with a different version of `protoc-gen-grpc-gateway` which caused a bunch of extra changes on the generated go files.